fix(std): parse_int return type consistency (i32→i64) across codegen, interpreter, type inference, stdlib, and tests
- src/core/infer/call/simple.rs: str_parse_int return type (bool,i32) → (bool,i64)
- src/interp/call.rs: parse_int/parse_float string method fallback returns
  Result::Err(string) on parse failure instead of propagating ?-panic
- std/strings.mimi: parse_int free function return type (bool,i32) → (bool,i64)
- tests/real_world/core_try_operator.mimi: Res Ok(i32) → Ok(i64);
  pattern literal 10 (i32) replaced with Ok(n) => if n == 10 ...
- tests/real_world/generate_suite.py: sync template with above

Unblocks real_world integration tests: real_world_multiple_std_modules
(uses std::strings::parse_int via module import) and real_world_cli_suite.

L1: dual_real_world_* pass
L2: cargo test --tests 2867/0/6

+++ src/core/infer/call/simple.rs
@@ -1188,7 +1188,7 @@ impl<'a> Checker<'a> {
                 }
                 return Type::Tuple(vec![
                     Type::Name("bool".into(), vec![]),
-                    Type::Name("i32".into(), vec![]),
+                    Type::Name("i64".into(), vec![]),
                 ]);
             }
             "str_parse_float" => {
